Add open weather api

+++ infra/src/main/java/com/nexters/gaetteok/weather/model/MainField.java
@@ -5,6 +5,6 @@ import lombok.Getter;
 @Getter
 public class MainField {
 
-    private Integer temp;
+    private Double temp;
 
 }
